requirements.txt
streamlit
transformers==4.24.0
torch==2.0.1
pdf2image==1.16.0
Pillow==9.3.0
tokenizers==0.12.1